remove implicit $getEditor
