+++ packages/lexical-builder-devtools-core/src/BuilderGraphPlan.tsx
@@ -1,10 +1,5 @@
-import {
-  LexicalBuilder,
-  definePlan,
-  provideOutput,
-} from "@etrepum/lexical-builder";
+import { definePlan, provideOutput } from "@etrepum/lexical-builder";
 import { ReactPlan } from "@etrepum/lexical-react-plan";
-import invariant from "./shared/invariant";
 import { Suspense, lazy } from "react";
 
 export const BuilderGraphComponent = lazy(
@@ -15,11 +10,6 @@ export const BuilderGraphPlan = definePlan({
   name: "@etrepum/lexical-builder-devtools-core/BuilderGraphPlan",
   dependencies: [ReactPlan],
   register(editor) {
-    const builder = LexicalBuilder.fromEditor(editor);
-    invariant(
-      builder !== undefined,
-      "BuilderGraphPlan: Could not get a LexicalBuilder reference",
-    );
     return provideOutput({
       Component: () => (
         <Suspense>
